Fixed and Changed AimLine shader. Added a window glass part to the windows Room 03
Finally fixed the Aim Line shader fade in and out, Made 3 separate materials of that shader, so each potion has it's own line color. Also changed the shader so it's able to glow.

Futhermore, I've added a plane to the windows of Room 03, so players can't jump through the window. Gave this plane a decorative glass texture, so it makes sense why you can't jump through it, and it won't give a "invisible wall" feeling